添加评论功能并展示

+++ src/main/java/life/bokchoy/community/controller/PublishController.java
@@ -32,7 +32,7 @@ public class PublishController {
 
     @GetMapping("/publish/{id}")//前端class为publish method为post
     public String edit(
-            @PathVariable("id") Integer id,
+            @PathVariable("id") Long id,
             Model model){
 
         QuestionDTO questionDTO=questionService.getQuestionDTOByQuestionId(id);
@@ -49,7 +49,7 @@ public class PublishController {
             @RequestParam(value = "title",required = false) String title,
             @RequestParam(value = "description",required = false) String description,
             @RequestParam(value = "tag",required = false) String tag,
-            @RequestParam(value = "id",required = false) Integer id,
+            @RequestParam(value = "id",required = false) Long id,
             HttpServletRequest request,
             Model model){
